+++ library/schemas/__init__.py
@@ -1,9 +1,12 @@
 """Convenient exports for schema definitions."""
 
 from .activities import ActivitiesSchema, configure_activity_schema
+from .activity import ActivitySchema
+from .assay import AssaySchema
 from .assay_postprocessing import AssayPostprocessSchema
 from .assays import AssaysSchema
 from .celllines import CellLinesSchema
+from .document import DocumentSchema
 from .documents import DocumentsSchema
 from .meta import CsvMetaSchema
 from .normalize import (
@@ -15,18 +18,25 @@ from .normalize import (
     normalize_testitems,
     normalize_tissues,
 )
+from .target import TargetSchema
 from .targets import TargetsSchema
+from .testitem import TestitemSchema
 from .testitems import TestitemsSchema
 from .tissues import TissuesSchema
 
 __all__ = [
     "ActivitiesSchema",
+    "AssaySchema",
     "AssaysSchema",
     "AssayPostprocessSchema",
+    "ActivitySchema",
     "DocumentsSchema",
+    "DocumentSchema",
     "CellLinesSchema",
     "TissuesSchema",
+    "TargetSchema",
     "TargetsSchema",
+    "TestitemSchema",
     "TestitemsSchema",
     "CsvMetaSchema",
     "normalize_activities",